# Conflicts:
#	src/main/java/com/zoomers/GameSetMatch/controller/UserController.java
#	src/main/java/com/zoomers/GameSetMatch/entity/User.java
#	target/classes/com/zoomers/GameSetMatch/controller/UserController.class
#	target/classes/com/zoomers/GameSetMatch/entity/User.class

+++ src/main/java/com/zoomers/GameSetMatch/controller/Tournament/RequestBody/IncomingRegistration.java
@@ -3,11 +3,11 @@ package com.zoomers.GameSetMatch.controller.Tournament.RequestBody;
 import java.util.List;
 
 public class IncomingRegistration {
-    private Long userID;
+    private Integer userID;
     private List<IncomingAvailability> availabilities;
     private String skillLevel;
 
-    public Long getUserID() {
+    public Integer getUserID() {
         return userID;
     }
